+++ .php-cs-fixer.php
@@ -151,6 +151,7 @@ $finder = Finder::create()
     ->in([
         __DIR__ . '/src',
         __DIR__ . '/config',
+        __DIR__ . '/dev',
         __DIR__ . '/tests',
     ])
     ->name('*.php')